Add retry logic to watch requests
This starts using the client-go RetryWatcher API which should handle
almost all network flakes seamlessly on the watch requests. One major
exception is when the last resourceVersion on the list request that the
RetryWatcher API is aware of no longer exists on the Kubernetes cluster.
A new watch request must be created in that case from the latest
resourceVersion.

Alternatively, the informer API could have been used, but this would
involve caching the actual objects, which is not desired for scale.

Relates:
https://issues.redhat.com/browse/ACM-2449

Signed-off-by: mprahl <mprahl@users.noreply.github.com>