Add ability to watch a selector
Instead of just a named object, this allows a user to watch a list with
a label selector in the same kind of API. The new selector field in an
ObjectIdentifier is the string representation of a LabelSelector.

This implementation only de-duplicates watches with the exact same
selector string - there might be multiple equivalent representations for
a selector, and the library does not attempt to identify these. It also
does not de-duplicate watches or events when a watcher watches both a
selector and a specifically named object that might be covered by that
selector.

Refs:
 - https://issues.redhat.com/browse/ACM-6428

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>