DeepCopy by default from the cache
Not DeepCopying can save some resources in some situations, but is also
a common way to introduce hard-to-find bugs, since items in the cache
can then mutate unexpectedly. This library now follows the pattern from
controller-runtime, which does a DeepCopy by default, which can be
disabled with an option when creating the client.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>