Add NewWithClients function
This allows a consumer of the package to not store/use a rest.Config to
create a DynamicWatcher, if they already have dynamic and discovery
clients.

Internally, this also changes the stored DiscoveryClient in the object
cache to use an interface, rather than a specific struct, which helps
provide some flexibility to users (for example, for testing).

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>