Return unwatchable resources with an error
Before v0.10.1, GVKtoGVR would return the first matching kind for the
resource, even if that resource did not support the `watch` API. In
v0.10.1, those resources were ignored, likely resulting in a "not found"
error being returned.

Now, when no matching resource that supports `watch` is found, it will
return the first resource found which matches the kind, along with a
special new error: `ErrResourceUnwatchable`. Returning the first
matching resource aligns with the "normal" behavior of returning early
when a resource does support `watch`, and seems to work better than
returning the last matching resource, which is often be a subresource.

This will help users that might want GVKtoGVR without watching an
object, and they can specially handle this error when it occurs.

Refs:
 - https://issues.redhat.com/browse/ACM-19965

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>